Compare paths of testfiles with FileSystem.arePathsSame
Fixes #8627. When the path of an actual test file (given by
vscode.Uri.fsPath) is later compared against the fullPath of a single
item of tests.TestFiles the drive letter mismatches on Windows and
aborts further actions despite the rest of the path is the same. As a
result code lenses on parametrized tests do not open a dropdown for
selecting single tests.